requirements.txt
# Main dependencies
pywin32>=228
comtypes>=1.1.14
pytest>=7.3.1
pytest-cov>=4.1.0 
fastmcp>=0.1.0
flask
google-generativeai
PySide6